+++ src/containers/Form/Users/layout.js
@@ -113,7 +113,7 @@ export default function UsersForm({
             id="frm-users-no-password-id"
             disabled={isSaving || !userHasUpdatePermission || currentAgentId === initialValues.get('id')}
             options={[
-              { label: 'Use Tenant Default: Enabled', value: 'null' },
+              { label: 'Use Tenant Default', value: 'null' },
               { label: 'Enabled', value: false },
               { label: 'Disabled', value: true }
             ].filter(val => (currentAgentId === initialValues.get('id') ? val.value !== 'null' : val))}